(PC-30077) refacto(VenueListModule): refacto and remove logic for the web
